Generic LCD Function Bug

+++ mycodo/functions/display_ssd1306_oled_128x32_spi.py
@@ -316,7 +316,7 @@ FUNCTION_INFORMATION = {
         },
         {
             'id': 'measure_max_age',
-            'type': 'float',
+            'type': 'integer',
             'default_value': 360,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
